Move GitHub tools from /actions to /tools directory
